first impl of superpages

+++ kernel/string.c
@@ -1,10 +1,10 @@
 #include "types.h"
 
 void*
-memset(void *dst, int c, uint n)
+memset(void *dst, int c, uint64 n)
 {
   char *cdst = (char *) dst;
-  int i;
+  uint64 i;
   for(i = 0; i < n; i++){
     cdst[i] = c;
   }
